requirements.txt
fastapi
uvicorn
pillow
supabase
pydantic
python-dotenv

# ML
pandas
scikit-learn
joblib
